chore: tidy admin user search header
